run_* functions; not parsed bocs; fees
