+++ ton_client/client/src/json_interface/modules.rs
@@ -313,15 +313,18 @@ pub struct TvmModule;
 
 fn register_tvm(handlers: &mut RuntimeHandlers) {
     let mut module = ModuleReg::new::<TvmModule>(handlers);
-    module.register_type::<crate::tvm::execute_message::ExecutionMode>();
-    module.register_type::<crate::tvm::execute_message::ExecutionOptions>();
+    module.register_type::<crate::tvm::types::ExecutionOptions>();
     module.register_async_fn(
-        crate::tvm::execute_message,
-        crate::tvm::execute_message::execute_message_api,
+        crate::tvm::run_executor,
+        crate::tvm::run_message::run_executor_api,
     );
     module.register_async_fn(
-        crate::tvm::execute_get,
-        crate::tvm::execute_get::execute_get_api,
+        crate::tvm::run_tvm,
+        crate::tvm::run_message::run_tvm_api,
+    );
+    module.register_async_fn(
+        crate::tvm::run_get,
+        crate::tvm::run_get::run_get_api,
     );
     module.register();
 }
