+++ ton_client/client/src/tests/common.rs
@@ -3,7 +3,7 @@ use crate::net::{ParamsOfQueryCollection, ResultOfQueryCollection};
 
 #[test]
 fn test_parallel_requests() {
-    let client1 = TestClient::new();
+    let client1 = std::sync::Arc::new(TestClient::new());
     let client2 = TestClient::new();
     let client3 = client1.clone();
