+++ ton_client/client/src/tvm/tests.rs
@@ -13,14 +13,13 @@
  */
 
 use crate::abi::encode_account::{ParamsOfEncodeAccount, StateInitSource};
-use crate::abi::{MessageSource, CallSet, DeploySet, ParamsOfEncodeMessage, Signer};
-use crate::boc::required_boc;
+use crate::abi::{Abi, CallSet, DeploySet, ParamsOfEncodeMessage, ResultOfEncodeMessage, Signer};
 use crate::json_interface::modules::{AbiModule, TvmModule};
 use crate::tests::{TestClient, SUBSCRIBE};
-use crate::tvm::execute_message::ExecutionMode;
-use crate::tvm::{execute_get, execute_message, ParamsOfExecuteGet, ParamsOfExecuteMessage};
+use super::*;
 use api_info::ApiModule;
 use serde_json::Value;
+use std::sync::Arc;
 
 const ELECTOR_ADDRESS: &str = "-1:3333333333333333333333333333333333333333333333333333333333333333";
 const ELECTOR_CODE: &str = "te6ccgECXgEAD04AART/APSkE/S88sgLAQIBIAMCAFGl//8YdqJoegJ6AhE3Sqz4FXkgTio4EPgS+SAs+BR5IHF4E3kgeBSYQAIBSBcEEgGvDuDKmc/+c4wU4tUC3b34gbdFp4dI3KGnJ9xALfcqyQAGIAoFAgEgCQYCAVgIBwAzs+A7UTQ9AQx9AQwgwf0Dm+hk/oAMJIwcOKABbbCle1E0PQFIG6SMG3g2zwQJl8GbYT/jhsigwf0fm+lIJ0C+gAwUhBvAlADbwICkTLiAbPmMDGBUAUm5h12zwQNV8Fgx9tjhRREoAg9H5vpTIhlVIDbwIC3gGzEuZsIYXQIBIBALAgJyDQwBQqss7UTQ9AUgbpJbcODbPBAmXwaDB/QOb6GT+gAwkjBw4lQCAWoPDgGHuq7UTQ9AUgbpgwcFRwAG1TEeDbPG2E/44nJIMH9H5vpSCOGAL6ANMfMdMf0//T/9FvBFIQbwJQA28CApEy4gGz5jAzhUACO4ftRND0BSBukjBwlNDXCx/igCASAUEQIBWBMSAl+vS22eCBqvgsGPtsdPqIlAEHo/N9KQR0cBbZ43g6kIN4EoAbeBAUiZcQDZiXM2EMBdWwInrA6A7Z5Bg/oHN9DHQW2eSRg28UAWFQJTtkhbZ5Cf7bHTqiJQYP6PzfSkEdGAW2eKQg3gSgBt4EBSJlxANmJczYQwFhUCSts8bYMfjhIkgBD0fm+lMiGVUgNvAgLeAbPmMDMD0Ns8bwgDbwREQQIo2zwQNV8FgCD0Dm+hkjBt4ds8bGFdWwICxRkYASqqgjGCEE5Db2SCEM5Db2RZcIBA2zxWAgHJMRoSAW4a85Q1ufW1LEXymEEC7IZbucuD3mjLjoAesLeX8QB6AAhIIRsCAUgdHAHdQxgCT4M26SW3Dhcfgz0NcL//go+kQBpAK9sZJbcOCAIvgzIG6TXwNw4PANMDIC0IAo1yHXCx/4I1EToVy5k18GcOBcocE8kTGRMOKAEfgz0PoAMAOgUgKhcG0QNBAjcHDbPMj0APQAAc8Wye1Uf4UwIBIB8eA3k2zx/jzIkgCD0fG+lII8jAtMfMPgju1MUvbCPFTFUFUTbPBSgVHYTVHNY2zwDUFRwAd6RMuIBs+ZsYW6zgXUhcA5MAds8bFGTXwNw4QL0BFExgCD0Dm+hk18EcOGAQNch1wv/gCL4MyHbPIAk+DNY2zyxjhNwyMoAEvQA9AABzxbJ7VTwJjB/4F8DcIFQgIAAYIW6SW3CVAfkAAbriAgEgMCICASAlIwOnTbPIAi+DP5AFMBupNfB3DgIo4vUySAIPQOb6GOINMfMSDTH9P/MFAEuvK5+CNQA6DIyx9YzxZABIAg9EMCkxNfA+KSbCHif4rmIG6SMHDeAds8f4XSRcAJYjgCD0fG+lII48AtM/0/9TFbqOLjQD9AT6APoAKKsCUZmhUCmgBMjLPxbL/xL0AAH6AgH6AljPFlQgBYAg9EMDcAGSXwPikTLiAbMCASApJgP1AHbPDT4IyW5k18IcOBw+DNulF8I8CLggBH4M9D6APoA+gDTH9FTYbmUXwzwIuAElF8L8CLgBpNfCnDgIxBJUTJQd/AkIMAAILMrBhBbEEoQOU3d2zwjjhAxbFLI9AD0AAHPFsntVPAi4fANMvgjAaCmxCm2CYAQ+DPQgVFMnArqAENch1wsPUnC2CFMToIASyMsHUjDLH8sfGMsPF8sPGss/E/QAyXD4M9DXC/9TGNs8CfQEUFOgKKAJ+QAQSRA4QGVwbds8QDWAIPRDA8j0ABL0ABL0AAHPFsntVH8oWgBGghBOVlNUcIIAxP/IyxAVy/+DHfoCFMtqE8sfEss/zMlx+wAD9yAEPgz0NMP0w8x0w/RcbYJcG1/jkEpgwf0fG+lII4yAvoA0x/TH9P/0//RA6MEyMt/FMofUkDL/8nQURq2CMjLHxPL/8v/QBSBAaD0QQOkQxORMuIBs+YwNFi2CFMBuZdfB21wbVMR4G2K5jM0pVySbxHkcCCK5jY2WyKAvLSoBXsAAUkO5ErGXXwRtcG1TEeBTAaWSbxHkbxBvEHBTAG1tiuY0NDQ2UlW68rFQREMTKwH+Bm8iAW8kUx2DB/QOb6HyvfoAMdM/MdcL/1OcuY5dUTqoqw9SQLYIUUShJKo7LqkEUZWgUYmgghCOgSeKI5KAc5KAU+LIywfLH1JAy/9SoMs/I5QTy/8CkTPiVCKogBD0Q3AkyMv/Gss/UAX6AhjKAEAagwf0QwgQRRMUkmwx4iwBIiGOhUwA2zwKkVviBKQkbhUXSwFIAm8iAW8QBKRTSL6OkFRlBts8UwK8lGwiIgKRMOKRNOJTNr4TLgA0cAKOEwJvIiFvEAJvESSoqw8StggSoFjkMDEAZAOBAaD0km+lII4hAdN/URm2CAHTHzHXC/8D0x/T/zHXC/9BMBRvBFAFbwIEkmwh4rMUAANpwhIB6YZp0CmGybF0xQ4xcJ/WJasNDpUScmQJHtHvtlFfVnQACSA3MgTjpwF9IgDSSa+Bv/AQ67JBg19Jr4G+8G2eCBqvgoFpj6mJwBB6BzfQya+DP3CQa4WP/BHQkGCAya+DvnARbZ42ERn8Ee2eBcGF/KGZQYTQLFQA0wEoBdQNUCgD1CgEUBBBjtAoBlzJr4W98CoKAaoc25PAXUE2MwSk2zzJAts8UbODB/QOb6GUXw6A+uGBAUDXIfoAMFIIqbQfGaBSB7yUXwyA+eBRW7uUXwuA+OBtcFMHVSDbPAb5AEYJgwf0U5RfCoD34UZQEDcQJzVbQzQDIts8AoAg9EPbPDMQRRA0WNs8Wl1cADSAvMjKBxjL/xbMFMsfEssHy/8B+gIB+gLLHwA8gA34MyBuljCDI3GDCJ/Q0wcBwBryifoA+gD6ANHiAgEgOTgAHbsAH/BnoaQ/pD+kP64UPwR/2A6GmBgLjYSS+B8H0gGBDjgEdCGIDtnnAA6Y+Q4ABHQi2A7Z5waZ+RQQgnObol3UdCmQgR7Z5wEUEII7K6El1FdXTjoUeju2wtfKSxXibKZ8Z1s63gQ/coRQXeBsJHrAnPPrB7PzAAaOhDQT2zzgIoIQTkNvZLqPGDRUUkTbPJaCEM5Db2SShB/iQDNwgEDbPOAighDudk9LuiOCEO52T2+6UhCxTUxWOwSWjoYzNEMA2zzgMCKCEFJnQ3C6jqZUQxXwHoBAIaMiwv+XW3T7AnCDBpEy4gGCEPJnY1CgA0REcAHbPOA0IYIQVnRDcLrjAjMggx6wR1Y9PAEcjomEH0AzcIBA2zzhXwNWA6IDgwjXGCDTH9MP0x/T/9EDghBWdENQuvKlIds8MNMHgCCzErDAU/Kp0x8BghCOgSeKuvKp0//TPzBFZvkR8qJVAts8ghDWdFJAoEAzcIBA2zxFPlYEUNs8U5OAIPQOb6E7CpNfCn7hCds8NFtsIkk3GNs8MiHBAZMYXwjgIG5dW0I/AiqSMDSOiUNQ2zwxFaBQROJFE0RG2zxAXAKa0Ns8NDQ0U0WDB/QOb6GTXwZw4dP/0z/6ANIA0VIWqbQfFqBSULYIUVWhAsjL/8s/AfoCEsoAQEWDB/RDI6sCAqoCErYIUTOhREPbPFlBSwAu0gcBwLzyidP/1NMf0wfT//oA+gDTH9EDvlMjgwf0Dm+hlF8EbX/h2zwwAfkAAts8UxW9mV8DbQJzqdQAApI0NOJTUIAQ9A5voTGUXwdtcOD4I8jLH0BmgBD0Q1QgBKFRM7IkUDME2zxANIMH9EMBwv+TMW1x4AFyRkRDAByALcjLBxTMEvQAy//KPwAe0wcBwC3yidT0BNP/0j/RARjbPDJZgBD0Dm+hMAFGACyAIvgzINDTBwHAEvKogGDXIdM/9ATRAqAyAvpEcPgz0NcL/+1E0PQEBKRavbEhbrGSXwTg2zxsUVIVvQSzFLGSXwPg+AABkVuOnfQE9AT6AEM02zxwyMoAE/QA9ABZoPoCAc8Wye1U4lRIA0QBgCD0Zm+hkjBw4ds8MGwzIMIAjoQQNNs8joUwECPbPOISW0pJAXJwIH+OrSSDB/R8b6Ugjp4C0//TPzH6ANIA0ZQxUTOgjodUGIjbPAcD4lBDoAORMuIBs+YwMwG68rtLAZhwUwB/jrcmgwf0fG+lII6oAtP/0z8x+gDSANGUMVEzoI6RVHcIqYRRZqBSF6BLsNs8CQPiUFOgBJEy4gGz5jA1A7pTIbuw8rsSoAGhSwAyUxKDB/QOb6GU+gAwoJEw4sgB+gICgwf0QwBucPgzIG6TXwRw4NDXC/8j+kQBpAK9sZNfA3Dg+AAB1CH7BCDHAJJfBJwB0O0e7VMB8QaC8gDifwLWMSH6RAGkjo4wghD////+QBNwgEDbPODtRND0BPQEUDODB/Rmb6GOj18EghD////+QBNwgEDbPOE2BfoA0QHI9AAV9AABzxbJ7VSCEPlvcyRwgBjIywVQBM8WUAT6AhLLahLLH8s/yYBA+wBWVhTEphKDVdBJFPEW0/xcbn16xYfvSOeP/puknaDtlqylDccABSP6RO1E0PQEIW4EpBSxjocQNV8FcNs84ATT/9Mf0x/T/9QB0IMI1xkB0YIQZUxQdMjLH1JAyx9SMMsfUmDL/1Igy//J0FEV+RGOhxBoXwhx2zzhIYMPuY6HEGhfCHbbPOAHVVVVTwRW2zwxDYIQO5rKAKEgqgsjuY6HEL1fDXLbPOBRIqBRdb2OhxCsXwxz2zzgDFRVVVAEwI6HEJtfC3DbPOBTa4MH9A5voSCfMPoAWaAB0z8x0/8wUoC9kTHijocQm18LdNs84FMBuY6HEJtfC3XbPOAg8qz4APgjyFj6AssfFMsfFsv/GMv/QDiDB/RDEEVBMBZwcFVVVVECJts8yPQAWM8Wye1UII6DcNs84FtTUgEgghDzdEhMWYIQO5rKAHLbPFYAKgbIyx8Vyx9QA/oCAfoC9ADKAMoAyQAg0NMf0x/6APoA9ATSANIA0QEYghDub0VMWXCAQNs8VgBEcIAYyMsFUAfPFlj6AhXLahPLH8s/IcL/kssfkTHiyQH7AARU2zwH+kQBpLEhwACxjogFoBA1VRLbPOBTAoAg9A5voZQwBaAB4w0QNUFDXVxZWAEE2zxcAiDbPAygVQUL2zxUIFOAIPRDW1oAKAbIyx8Vyx8Ty//0AAH6AgH6AvQAAB7TH9Mf0//0BPoA+gD0BNEAKAXI9AAU9AAS9AAB+gLLH8v/ye1UACDtRND0BPQE9AT6ANMf0//R";
@@ -30,10 +29,10 @@ const ELECTOR_DATA: &str = "te6cckICAdwAAQAAXWYAAANP5zNFdL1WHOmhM8muxAeRTL3uvNJv
 async fn test_execute_get() {
     TestClient::init_log();
     let client = TestClient::new();
-    let execute_get = client.wrap_async(
-        execute_get,
+    let run_get = client.wrap_async(
+        run_get,
         TvmModule::api(),
-        crate::tvm::execute_get::execute_get_api(),
+        crate::tvm::run_get::run_get_api(),
     );
     let encode_account = client.wrap_async(
         crate::abi::encode_account::encode_account,
@@ -55,8 +54,8 @@ async fn test_execute_get() {
         .await
         .account;
 
-    let result = execute_get
-        .call(ParamsOfExecuteGet {
+    let result = run_get
+        .call(ParamsOfRunGet {
             account: elector.clone(),
             function_name: "participant_list".into(),
             input: None,
@@ -69,8 +68,8 @@ async fn test_execute_get() {
         r#"[[["0x101b6d65a384b9c70deb49fd6c43ffc0f60ed22fcc3a4966f7043794a749228","0x36b1f820a400"],[["0x3de5d8590fe6ad191bf94d4136dfb630e9b3447bb2f1a6ae2d8e3e4cbee1d9f","0x377aab54d000"],[["0x558f90c0682d677b46005ce2e04206c255ea9a05bfac0ff5aea9d7182a28913","0x36b1f820a400"],[["0x7698228973a595751d79e1fafd5a4145b3d35349bf0b43322afb61b138f01eb","0x36b1f820a400"],[["0x9d1ef8a40a9fbf1ca505f072258048ec15e0637baa085649d77b9a90220003e","0x36b1f820a400"],[["0xac21ef27c8ed4487270f1c45e99dac091ca4007951217ece344452df7047e5a","0x36b1f820a400"],[["0xaed529418ab67a31a4b98c224f8fdb2fec11f0100c23751e67b312cba11fb23","0x36b1f820a400"],[["0xbd14cdade9067c523f44fd208dee5daa7d852151725d713b92c840a031018a6","0x36b1f820a400"],[["0xbddff0d98f42a3155e577a5579623a911e3b03401835166553f88375cbd9657","0x36b1f820a400"],[["0x12893bbd649bf2e1e79cb084025638cdd7906eebca40efeee7fdbd548cc96391","0x36b1f820a400"],[["0x15546bc7b5124f6d83d6c5a62b8890a48b933168e141c01229431a6c0c499780","0x36b1f820a400"],[["0x1cbea6a399ba200958db255579cda2195006f3a3108b2d6ef7e258e42c101479","0x36b1f820a400"],[["0x1f8ee6ba2902715804c769c3845b6b3a37802e462e8df63ba19f827a92dbbda0","0x36b1f820a400"],[["0x1fdd556d84d1d9f24a739c2600ec72256cc00920d85ad3a2edb3e0d72146789d","0x36b1f820a400"],[["0x20f20c2cfa4d72afb9c5f64d4735070962b3323b3629892c75f56427f175ebe4","0x36b1f820a400"],[["0x219d32737b0f3769869b8fa750ba8e3cd9f19b21a4d669c7c79d420d7f7cdb24","0x36b1f820a400"],[["0x2615b4aeb69140531228248a9d84593117b64e22d462e3968e39c1840a260523","0x36b1f820a400"],[["0x26984c9f04bc1889061e98bd9caf6955f750219d8e8dbc0986feb9d770e5a15b","0x36b1f820a400"],[["0x28bb07d80e20aa624ae47dfe53f915b23a666bf825ff283bac06c14bea1eaa74","0x36b1f820a400"],[["0x2a23566008fd4f87105b09d02c739452e45187fbada5e7e52ada356264cd6751","0x36b1f820a400"],[["0x2dcc70859876106b21b598ba9a10c9932259c36f44adeb95a178e67f6afd2f7a","0x36b1f820a400"],[["0x30b854226ef943d738d2dfcc72ede3b39d08604ca7211abf3c76f488441c77fc","0x36b1f820a400"],[["0x31bb74a5a53769d3db789d961375ea569d4da0bd6ac2b12f830dd6be81968ef1","0x36b1f820a400"],[["0x334f22e0de2e24a070fec7c1d77d7a988a79d66b79e2d654310a963964edd337","0x36b1f820a400"],[["0x36c44eddf773390cdb42f93f8454ea9c7ca45aa8948346df8f642a59ce44c442","0x36b1f820a400"],[["0x3d29d2b5ceef46703255ce8cfe3aad3c4fefd3a2025e5a48ce78a63f20887eff","0x36b1f820a400"],[["0x41b047a20ed691e9376f7f2f60d6571290e34ef4e1b85467dcc3d7c0cf7fae90","0x36b1f820a400"],[["0x41e7541c377b58a0cfc4ca954731e971f6dc9fa6806eaa1709d011d3d32593ce","0x36b1f820a400"],[["0x426a52d3b3d016451c46b3a0eacb382fbbb38739e00d041d4038f795a54e25b5","0x36b1f820a400"],[["0x42f89915ca540af691f623f201b616caa7f5e104f8293698f8b46c4e7bb5b292","0x36b1f820a400"],[["0x4449521e793b02b036ce698c3af951e9548cd5b862b704fa5cc9e80b171a3c61","0x36b1f820a400"],[["0x492f4fee6a035a09e9ff09d65a65768899b04797cff08dc2c64ae11cd94d1968","0x36b1f820a400"],[["0x4947018f9c0c9302b2783eb5edacb76ccae3b5c5a2f6355b5b51afd1a18075f8","0x36b1f820a400"],[["0x4c27708e4ce81a0bbeb315ece024ba495f3e3fab5f83a2941b7731a58ad32160","0x36b1f820a400"],[["0x5059d40f80f578c3c384239415f54af35ab4dbdea0251618d4c3c7b4937e7e69","0x36b1f820a400"],[["0x5191f8cbfe1ce25a68c337ede75638321374112b868584092a335f83caad59a4","0x36b1f820a400"],[["0x531296c32ea64d09dcb44ff0b99843dc9855143c70b9fad42deb33881525fb84","0x36b1f820a400"],[["0x54c9860aa34ddba2a16e4f4271e1771f61f1e8a7a116fbbfa62f0e535b95559e","0x36b1f820a400"],[["0x54ce2d6b35d0d670e37fcd533ff17c2116e0acead719194e46d478944b33108e","0x36b1f820a400"],[["0x576af5e4af963a0caa957629d009906119e418e7f7778f5a55d41c0905b73a4f","0x36b1f820a400"],[["0x59905476f4781f6a79359079bfa3fe295c65d6b918afadbc352edcdb558ad094","0x36b1f820a400"],[["0x5a4e95cdf94bed240ebabded084b70b2548601686d94a751f240aedd2032e4f2","0x36b1f820a400"],[["0x5a7500f11becf6741fe5624d2298f6b830ead261871a48a81f80bf9be09ed866","0xa3b5840f4000"],[["0x5c26942bf33c49485db3b2693e5d582708b44705f712c4e24af1ee84744c079e","0x36b1f820a400"],[["0x5fcdcc107e81ef4399c9d603a25fcba75cb78f1fa1bafd3acb39e3521d7fc9ce","0x36b1f820a400"],[["0x6107f5b2974fabf6f0aa1a7898340b3f76c4ba272b95a3e4bb809c1d529b6997","0x36b1f820a400"],[["0x647b9a476f733ec5ee9cbc0bfb021335cd3166b9aaa8ad27ed0f88d9f6bf9dbe","0x36b1f820a400"],[["0x658c461d8dad54a5a9cbbbb2711920a541ba58003e7029cc228dfdfbc17ede3f","0x36b1f820a400"],[["0x661336351b889e0124fbc19f9f35f6a0f6e8c4fa9b89e9ef527718bd6aa254be","0x36b1f820a400"],[["0x6852746bbfb41e556daae99d375b2839ad62b35355c3b9fbbd54b4946ae2050f","0x36b1f820a400"],[["0x68ad3d98642913848b605dbad3f1df971f21908d360c37af4a493e9b4646b45e","0x36b1f820a400"],[["0x6c07c6be93940a83b30514b21531fd3dd204bb89e7f77b5a2421a41d4e85c74d","0x36b1f820a400"],[["0x6d4ad504054f292b7f66c7ed32f3b123bfa5c7be9c45faf26d77ad85efa64a38","0x36b1f820a400"],[["0x6e77d45d07651565be5cdb11b80c91fa18def0a434f246c0b25bb50fc4877dd4","0x36b1f820a400"],[["0x738600c570c19ef1b91bf2cd83709d71899c246bfabb5b08dc70fe32b5c81f7d","0x36b1f820a400"],[["0x7469b663b9fa7be185aa1819bfb48a4eda6e4d8af33e1955d95fa5e156d50f12","0x36b1f820a400"],[["0x77410e09363239b0999198a701e37f75775cc55049ba541497967f5d8ba74ef4","0x36b1f820a400"],[["0x781c96175cf45b791142326964347095fd0fbdd3c8579c42cea108798e025152","0x36b1f820a400"],[["0x79b43e9c18241636ae7c554097bb4bc5da03249bee67abe5366a5b093b708cab","0x36b1f820a400"],[["0x7ad807b91790868497768476e8c8e6b53ff9b1a91fbfe6a7edae8de0307a8157","0x36b1f820a400"],[["0x8308ff2b214d509d3781d7361a7ccb5f4fb976f8e386ce3c9082bcad8805d13d","0x36b1f820a400"],[["0x845a0fff44669c941475eb3f3ffd6e065ee94cfbfdcb820877744d6f9647a5d0","0x2d62042b6e00"],[["0x88700f083f3bc7971c348de8357ee36b2551d8cdb7ea4b4e4e8aae558d67a231","0x36b1f820a400"],[["0x8b08c457cac18642f49ab7de0ef7551b93e11dbc2979062f22b271b890e8d2f0","0x36b1f820a400"],[["0x8bc840e0c5a98e608e70307ada41aa94a745a51f6065111942021a4e601dc328","0x36b1f820a400"],[["0x93e518529faa2244ee1bdc24a5459d4b3d2047f8756b12636e2cba3b766ec201","0x36b1f820a400"],[["0x993d90fac526bdad11549104105452e9198da8d485dbb4af17b044a721fa8b82","0x36b1f820a400"],[["0x9997880b1dcc011ce4fefeb587eca16c027c81aafeb4305d3a1755182c269b5e","0x36b1f820a400"],[["0x9d998de650f13c85da4ab08de0fa7960771d4269081fa1ed1f9940c5cd8bb57c","0x36b1f820a400"],[["0x9fab138505d28c3c2d68509c5414abe933ab7de90610d8cc84edaf380e739f48","0x36b1f820a400"],[["0x9fd585f4d71c50ff54b69255ddbaa4a30eae31cde2d02ba6d4c0f87faf288f9a","0x36b1f820a400"],[["0xa42d598e3d6c051880488bfd139705c9853ff2e93046c6e096eba5f5b8dd714e","0x36b1f820a400"],[["0xa6e3ff7b1f340f7d02a1b64ade185c9039cd2751ef47ac5d7950b527b377d566","0x36b1f820a400"],[["0xa79d52472a9343b4f91c61b7e065cd736844064c11188fb86fef32447b163462","0x36b1f820a400"],[["0xa82bdb918a99f7192b0ebe745f04217991d2077dc43ffe75956782f55c7c9805","0x36b1f820a400"],[["0xa87f60cfad2f10ec420d4660d98a43a1105a867aac63a2724075f155b991fd35","0x36b1f820a400"],[["0xad1e503c43f7f62bb672b234eb1510b8ccef4d23b6de1f53a8a0d738c961cee2","0x36b1f820a400"],[["0xad8dd15447ac5c3b0ac9ed9ebae3b32cfe3cda5442bfce7843443a353701eb34","0x36b1f820a400"],[["0xade82619842d2257fb19097c990b77818f2352e3809b9c179b3b66989b8e01c3","0x36b1f820a400"],[["0xb739a017b3b9c9577eeba0d3b94fe2027333ccdad378f0aa67b441eb8cbd675a","0x36b1f820a400"],[["0xb85ed5c5a48abadc5bf4a85185f781aa60eebe7ef20642f660c7e90d481984cf","0x36b1f820a400"],[["0xbcb7406f71b46a5171b822f609d50df1a485bbae832f76db0a356b243616ca8e","0x36b1f820a400"],[["0xbedc1da66f906866cf8af7e57cda645018c39d1b028e9ed4682643941c940348","0x36b1f820a400"],[["0xc2660177ec158c05676b396baab45f8f8a63f74a0eaf1a7cfe011c7eea0cd8a4","0x340141969800"],[["0xc536058376e87f6481ce31b0e088235b4be9df00145c97081c45a28cce64c684","0x36b1f820a400"],[["0xc8fd550fcf32a9ea6aef295e788e4394e744f0939ab5fa8b9009577e274a477e","0x36b1f820a400"],[["0xcbfe056a9e9fafd246a8fa3025c3d870dac6b01f68adc847f03277eac906452f","0x36b1f820a400"],[["0xd023735d89cb9d29c5301b87f00d3f7a42aa6f0320086473e50b7e3b8b9acb12","0x36b1f820a400"],[["0xd17002d5872d62876fc4cae771c472ececae0b50820d760718a753acf431f31c","0x36b1f820a400"],[["0xd847bac558e925bd87b15a9e8c077df36537e6fc52d5c2019004c0c570fd0266","0x36b1f820a400"],[["0xdab17536c875995ce144f17771c79e7e9d6adcaaa66cce64947c8d17a363a2dd","0x36b1f820a400"],[["0xdf0b5c031ece9dadfd23c63a41e4e7f1ae4138b157ff7588c21083853d585789","0x36b1f820a400"],[["0xe087dae3faaf4748c8bcd237ca7ece5f8bcddf6b60db216252c5e29a7f6a33a4","0x36b1f820a400"],[["0xe09755d90d62160409b67ff28584f2800087f9063d76b3240c4d3f94f0880c41","0x36b1f820a400"],[["0xe23c47f9c2e9d2d87d9f1fcc0352ef28d13f322f8003d4210bb33692e77fd988","0x36b1f820a400"],[["0xeb2269b0ea934046a59399bd824f6a7fae4c7d696bb163e1bd235cbc21aa2b55","0x36b1f820a400"],[["0xedf8d203bebac7d629840ca0a704cbff92607d6bf538bc99ab65fada6a7b3c65","0x36b1f820a400"],[["0xf179ca30b1a9c8c33e33863b04d8d0078dddbf974a1e8666d072e0403c997f21","0x36b1f820a400"],[["0xf199c75a842c96c459272502b7010a78f29e7d00ad649f7756dd11c8d321a97c","0x36b1f820a400"],[["0xf19a880d58384bfaf5c839e7b6502ff7e6dc11cc38a77f651c948ea8475a37f7","0x36b1f820a400"],[["0xf25841168bb223f03cc01f5934474e56ae3ac0307a048ab167326c7d655c25db","0x4ef873b38200"],[["0xf25e305cc44404dae89a8f3b577cf94c367ab28a8ebc81a5c551d39303e254c2","0x36b1f820a400"],[["0xf3c0eed928d059ec9c99fd55ef4df9ccdaa30626e14b833f064936099b8088e1","0x36b1f820a400"],[["0xf6fe4424c9df211b1d2e92f7a91889aac643c605de458fa8f2af90534b885654","0x8cbf79329400"],[["0xf8a26375e76f1f8ff763787507fe2e01ed257b1a6b1772e48338606862a80da4","0x36b1f820a400"],[["0xfb471071aa87f25465da8c98bdeb1b24165b4a1694c5a6ab9f59eb57ce9e451d","0x36b1f820a400"],[["0xfb66f351a3b27e1702a21bfd189f3db5053f9f0089b26e7f05218fa87b925e2e","0x36b1f820a400"],[["0xfbcd15956e466a3c945c8bee6e6bc6bdab6b1b2ec0c07a3ba431091795751bef","0x36b1f820a400"],[["0xfced4379f1cb13157b34d50301a65ab47dc3452f4cd0e2a2d8e0b33a07350f43","0x36b1f820a400"],null]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]"#
     );
 
-    let result = execute_get
-        .call(ParamsOfExecuteGet {
+    let result = run_get
+        .call(ParamsOfRunGet {
             account: elector.clone(),
             function_name: "compute_returned_stake".into(),
             input: Some(Value::String(format!(
@@ -86,8 +85,8 @@ async fn test_execute_get() {
         .output;
     assert_eq!(result[0], "0x0");
 
-    let result = execute_get
-        .call(ParamsOfExecuteGet {
+    let result = run_get
+        .call(ParamsOfRunGet {
             account: elector.clone(),
             function_name: "past_elections".into(),
             input: None,
@@ -100,151 +99,148 @@ async fn test_execute_get() {
 }
 
 #[tokio::test(core_threads = 2)]
-async fn test_execute_message() {
+async fn test_run_executor() {
+    let run = |client: Arc<TestClient>, message: ResultOfEncodeMessage, abi: Abi, account: String| async move {
+        let run_executor = client.wrap_async(
+            run_executor,
+            TvmModule::api(),
+            run_message::run_executor_api(),
+        );
+        let result = run_executor
+            .call(ParamsOfRunExecutor {
+                message: message.message,
+                abi: Some(abi.clone()),
+                account: Some(account.clone()),
+                execution_options: None,
+            })
+            .await;
+        
+        assert_eq!(result.transaction["in_msg"], message.message_id);
+        assert!(result.fees.total_account_fees > 0);
+
+        result.account
+    };
+
+    test_run_message(run).await;
+}
+
+#[tokio::test(core_threads = 2)]
+async fn test_run_tvm() {
+    let run = |client: Arc<TestClient>, message: ResultOfEncodeMessage, abi: Abi, account: String| async move {
+        let run_tvm = client.wrap_async(
+            run_tvm,
+            TvmModule::api(),
+            run_message::run_tvm_api(),
+        );
+        let result = run_tvm
+            .call(ParamsOfRunTvm {
+                message: message.message,
+                abi: Some(abi.clone()),
+                account: account.clone(),
+                execution_options: None,
+            })
+            .await;
+
+        result.account
+    };
+
+    test_run_message(run).await;
+}
+
+async fn test_run_message<F>(run: impl FnOnce(Arc<TestClient>, ResultOfEncodeMessage, Abi, String) -> F)
+where
+    F: futures::Future<Output=String>
+{
     TestClient::init_log();
-    let client = TestClient::new();
-    let execute_message = client.wrap_async(
-        execute_message,
+    let client = Arc::new(TestClient::new());
+    let run_tvm = client.wrap_async(
+        run_tvm,
         TvmModule::api(),
-        crate::tvm::execute_message::execute_message_api(),
+        run_message::run_tvm_api(),
     );
 
-    for abi_version in 1..=2 {
-        let (abi, tvc) = TestClient::package(SUBSCRIBE, Some(abi_version));
-        let keys = client.generate_sign_keys();
+    let (abi, tvc) = TestClient::package(SUBSCRIBE, None);
+    let keys = client.generate_sign_keys();
 
-        let wallet_address = "0:2222222222222222222222222222222222222222222222222222222222222222";
+    let wallet_address = "0:2222222222222222222222222222222222222222222222222222222222222222";
 
-        let address = client
-            .deploy_with_giver_async(
-                ParamsOfEncodeMessage {
-                    abi: abi.clone(),
-                    deploy_set: DeploySet::some_with_tvc(tvc.clone()),
-                    call_set: CallSet::some_with_function_and_input(
-                        "constructor",
-                        json!({
-                            "wallet": wallet_address.to_string(),
-                        }),
-                    ),
-                    signer: Signer::Keys { keys: keys.clone() },
-                    processing_try_index: None,
-                    address: None,
-                },
-                None,
-            )
-            .await;
+    let address = client
+        .deploy_with_giver_async(
+            ParamsOfEncodeMessage {
+                abi: abi.clone(),
+                deploy_set: DeploySet::some_with_tvc(tvc.clone()),
+                call_set: CallSet::some_with_function_and_input(
+                    "constructor",
+                    json!({
+                        "wallet": wallet_address.to_string(),
+                    }),
+                ),
+                signer: Signer::Keys { keys: keys.clone() },
+                processing_try_index: None,
+                address: None,
+            },
+            None,
+        )
+        .await;
 
-        let account: String = client.fetch_account(&address).await["boc"]
-            .as_str()
-            .unwrap()
-            .into();
+    let account: String = client.fetch_account(&address).await["boc"]
+        .as_str()
+        .unwrap()
+        .into();
 
-        // TvmOnly
+    let subscribe_params = json!({
+        "subscriptionId": "0x1111111111111111111111111111111111111111111111111111111111111111",
+        "pubkey": "0x2222222222222222222222222222222222222222222222222222222222222222",
+        "to": "0:3333333333333333333333333333333333333333333333333333333333333333",
+        "value": "0x123",
+        "period": "0x456",
+    });
 
-        let subscribe_params = json!({
-            "subscriptionId": "0x1111111111111111111111111111111111111111111111111111111111111111",
-            "pubkey": "0x2222222222222222222222222222222222222222222222222222222222222222",
-            "to": "0:3333333333333333333333333333333333333333333333333333333333333333",
-            "value": "0x123",
-            "period": "0x456",
-        });
+    // crate subscription
+    let message = client.encode_message(ParamsOfEncodeMessage {
+            address: Some(address.clone()),
+            abi: abi.clone(),
+            call_set: CallSet::some_with_function_and_input(
+                "subscribe",
+                subscribe_params.clone(),
+            ),
+            signer: Signer::Keys { keys: keys.clone() },
+            deploy_set: None,
+            processing_try_index: None,
+        }).await;
 
-        let result = execute_message
-            .call(ParamsOfExecuteMessage {
-                message: MessageSource::EncodingParams(ParamsOfEncodeMessage {
-                    address: Some(address.clone()),
-                    abi: abi.clone(),
-                    call_set: CallSet::some_with_function_and_input(
-                        "subscribe",
-                        subscribe_params.clone(),
-                    ),
-                    signer: Signer::Keys { keys: keys.clone() },
-                    deploy_set: None,
-                    processing_try_index: None,
-                }),
-                mode: ExecutionMode::TvmOnly,
-                account: Some(account.clone()),
-                execution_options: None,
-            })
-            .await;
+    let account = run(client.clone(), message, abi.clone(), account).await;
 
-        let result = execute_message
-            .call(ParamsOfExecuteMessage {
-                account: Some(required_boc(&Some(result.account)).unwrap()),
-                message: MessageSource::EncodingParams(ParamsOfEncodeMessage {
-                    abi: abi.clone(),
-                    call_set: CallSet::some_with_function_and_input(
-                        "getSubscription",
-                        json!({
-                            "subscriptionId": subscribe_params["subscriptionId"].clone(),
-                        }),
-                    ),
-                    signer: Signer::Keys { keys: keys.clone() },
-                    address: Some(address.clone()),
-                    deploy_set: None,
-                    processing_try_index: None,
-                }),
-                mode: ExecutionMode::TvmOnly,
-                execution_options: None,
-            })
-            .await;
-        assert_eq!(
-            result.decoded.unwrap().output.unwrap()[if abi_version == 1 {
-                "subscription"
-            } else {
-                "value0"
-            }]["pubkey"],
-            subscribe_params["pubkey"]
-        );
+    // check subscription
 
-        // Full
+    let message = client.encode_message(ParamsOfEncodeMessage {
+        abi: abi.clone(),
+        call_set: CallSet::some_with_function_and_input(
+            "getSubscription",
+            json!({
+                "subscriptionId": subscribe_params["subscriptionId"].clone(),
+            }),
+        ),
+        signer: Signer::Keys { keys: keys.clone() },
+        address: Some(address.clone()),
+        deploy_set: None,
+        processing_try_index: None,
+    }).await;
 
-        let result = execute_message
-            .call(ParamsOfExecuteMessage {
-                message: MessageSource::EncodingParams(ParamsOfEncodeMessage {
-                    address: Some(address.clone()),
-                    abi: abi.clone(),
-                    call_set: CallSet::some_with_function_and_input(
-                        "subscribe",
-                        subscribe_params.clone(),
-                    ),
-                    signer: Signer::Keys { keys: keys.clone() },
-                    deploy_set: None,
-                    processing_try_index: None,
-                }),
-                mode: ExecutionMode::Full,
-                account: Some(account.clone()),
-                execution_options: None,
-            })
-            .await;
-
-        let result = execute_message
-            .call(ParamsOfExecuteMessage {
-                account: Some(required_boc(&Some(result.account)).unwrap()),
-                message: MessageSource::EncodingParams(ParamsOfEncodeMessage {
-                    abi: abi.clone(),
-                    call_set: CallSet::some_with_function_and_input(
-                        "getSubscription",
-                        json!({
-                            "subscriptionId": subscribe_params["subscriptionId"].clone(),
-                        }),
-                    ),
-                    signer: Signer::Keys { keys: keys.clone() },
-                    address: Some(address.clone()),
-                    deploy_set: None,
-                    processing_try_index: None,
-                }),
-                mode: ExecutionMode::TvmOnly,
-                execution_options: None,
-            })
-            .await;
-        assert_eq!(
-            result.decoded.unwrap().output.unwrap()[if abi_version == 1 {
-                "subscription"
-            } else {
-                "value0"
-            }]["pubkey"],
-            subscribe_params["pubkey"]
-        );
-    }
+    let result = run_tvm
+        .call(ParamsOfRunTvm {
+            abi: Some(abi.clone()),
+            account,
+            message: message.message,
+            execution_options: None,
+        })
+        .await;
+    assert_eq!(
+        result.decoded.unwrap().output.unwrap()[if TestClient::abi_version() == 1 {
+            "subscription"
+        } else {
+            "value0"
+        }]["pubkey"],
+        subscribe_params["pubkey"]
+    );
 }
